Port WindowsHelper's src/ tree, excluding AntiLock
Mirrors WindowsHelper's current architecture (gui/interface/platform/utils
split, HotkeyRegistry, backup/CAN/Modbus logging, macro and settings-binding
subsystems, etc.) onto WindowsAddon, merging the app entry point and
stripping every AntiLock touchpoint since this fork does not ship that
feature. Also fixes two vendored libs/ bugs (the isotp snprintf macro
collision with boost::system, an out-of-bounds bitarray read) and the
MSBuild project (missing WIN32_LEAN_AND_MEAN/NOMINMAX, missing PCH
overrides, missing Wlanapi/Urlmon/Wtsapi32 link deps) needed to build the
result. Removes the now-superseded gui/CanPanel/BitEditorDialog in favor
of the ported gui/BitFieldEditorDialog.

+++ src/ModbusRegisterGrouping.hpp
@@ -8,6 +8,7 @@
 #include <optional>
 #include <utility>
 #include <vector>
+#include "ModbusTypeTraits.hpp"
 
 struct ModbusRegisterRange { size_t offset = 0; size_t count = 0; };
 
@@ -67,8 +68,8 @@ std::optional<GroupedModbusRegisterReadResult> ReadGroupedModbusRegisters(
     GroupedModbusRegisterReadResult result;
     for(const auto& range : BuildContiguousModbusRegisterRanges(items, branch))
     {
-        if(range.offset > std::numeric_limits<uint16_t>::max() - base_offset ||
-            range.count > std::numeric_limits<uint16_t>::max())
+        if(range.offset > static_cast<size_t>(std::numeric_limits<uint16_t>::max()) - base_offset ||
+            range.count > static_cast<size_t>(std::numeric_limits<uint16_t>::max()))
             return std::nullopt;
         auto registers = read_registers(static_cast<uint16_t>(base_offset + range.offset),
             static_cast<uint16_t>(range.count));
@@ -91,8 +92,8 @@ std::optional<GroupedModbusBitReadResult> ReadGroupedModbusBits(
     GroupedModbusBitReadResult result;
     for(const auto& range : BuildContiguousModbusRegisterRanges(items, branch))
     {
-        if(range.offset > std::numeric_limits<uint16_t>::max() - base_offset ||
-            range.count > std::numeric_limits<uint16_t>::max())
+        if(range.offset > static_cast<size_t>(std::numeric_limits<uint16_t>::max()) - base_offset ||
+            range.count > static_cast<size_t>(std::numeric_limits<uint16_t>::max()))
         {
             ++result.failed_read_count;
             continue;
